Remove creation time when removing file
There's an edge case where the creation time entry for a log file is
removed, due to FileRegistry#deleteWriter removing it when
FileRegistry#deleteWriters is called in Uploader#uploadFiles and
the committed offset has changed after getting the lock.

If the partition doesn't receive any new messages (because of a
rebalance), the log file will never be deleted (or rather trimmed away)
and Secor will just keep generating warning messages.

By moving the removal from `mCreationTimes` to FileRegistry#deletePath
this scenario should not occur.