Make it possible to override LogFilePath impl.
This has two sides to it:

* To enable proper DelimitedTextFile{Reader,Writer} implementations.
* To enable the user to decide (partly) how filenames are to be
  constructed.

The changes regarding proper DelimitedTextFile{Reader,Writer}
implementations are due to the fact that the implementations need to
know which offsets are in which file, which in the case of the
SequenceFile ones are already built in.

The current implementations will not handle trimming of log files
correctly, but for the sake of having the integration tests passing in
this repo I’ve left the functionality as-is, so it’s up to each user
if they want to implement a working solution.

The filename of log files are partly "configurable" as some parts are
needed to guarantee uniqueness:

* Always include offset as last filename component
* Always include Kafka partition in filename