Optionally base upload criteria on oldest files
Previously Secor would consider a partition eligible for upload if the
“youngest” file was created X seconds ago, which makes for a rather
non-intuitive behaviour for us.
With these changes it’ll be possible to base the decision to upload files
if anyone of them is older than X seconds.

Yes, the naming is extremely horrible, but I couldn’t come up with
anything better right now.